Fix #1020: Be more careful about converting a float into an integer
If the value overflows then it will get truncated. So convert to 64
bit and then down to a 32 bit value.

Original issue only seen on rpi, not on 32 bit intel compiles.

+++ src/sccz80/plunge.c
@@ -392,14 +392,14 @@ void plnge2a(int (*heir)(LVALUE* lval), LVALUE* lval, LVALUE* lval2, void (*oper
 
             if ( lval2->is_const && (lval->val_type == KIND_INT || lval->val_type == KIND_CHAR || lval->val_type == KIND_LONG) ) {
                 doconstoper = 1;
-                const_val = (uint32_t)lval2->const_val;
+                const_val = (int32_t)(int64_t)lval2->const_val;
                 clearstage(before, 0);
                 force(rhs_val_type, lhs_val_type, lval->ltype->isunsigned, lval2->ltype->isunsigned, 1);
             }
             /* Handle the case that the constant was on the left */
             if ( lval1_wasconst && (lval2->val_type == KIND_INT || lval2->val_type == KIND_CHAR || lval2->val_type == KIND_LONG) ) {
                 doconstoper = 1;
-                const_val = (uint32_t)lval->const_val;
+                const_val = (int32_t)(int64_t)lval->const_val;
                 clearstage(before_constlval, 0);
                 force(lhs_val_type, rhs_val_type, lval2->ltype->isunsigned, lval2->ltype->isunsigned,1);
             }
